Decompose Primary Awareness presentation files #296
